Moved AppBar into a decorator. It usually is not part of the displayed page it seems like a good separation. Also we can later create decorators like withToolBar or isModal etc.
